Initial commit
This contains the original vim folder with the already installed plugins on it.
These are:

Pathogen
Fugitive
omnicppcomplete
ftp for c and c++ ==> maybe not used?
localrc ==> not being used, maybe badly configured
some color schemes
customized vimrc

Work is required to install existing plugins as submodules and clean up.